Refactor int to size_t.

+++ ngraph/core/reference/include/ngraph/runtime/reference/convolution_backprop_data.hpp
@@ -155,7 +155,7 @@ namespace ngraph
                         // strides is already applied (through 0's extension in input)
                         // padding = pads_begin + pads_end, formula below is using
                         // params.pad_begin/params.pads_end:
-                        int expected_dim = (input_shape[i + 2] - 1) - filters_shape[i + 2] +
+                        size_t expected_dim = (input_shape[i + 2] - 1) - filters_shape[i + 2] +
                                            params.pads_begin[i] + params.pads_end[i] + 2 +
                                            params.output_padding[i];
                         if (out_shape_3d[i] != expected_dim)
